requirements.txt
Django==1.10.5
django-betterforms==1.1.4
django-bootstrap3==8.2.1
django-countries==4.0
django-widget-tweaks==1.4.1
olefile==0.44
Pillow==4.0.0
six==1.10.0
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
djangorestframework==3.6.2